+++ CHANGELOG.md
@@ -6,6 +6,7 @@ GAMSPy 1.4.1
 - General
   - Fix implicit parameter validation bug.
   - Migrate GAMSPy CLI to Typer.
+  - Fix solver options bug in frozen solve.
 - Testing
   - Lower the number of dices in the interrupt test and put a time limit to the solve.
 - Documentation